+++ juntagrico/config.py
@@ -33,7 +33,8 @@ class Config:
             'member_type': _('Mitglied'),
             'member_type_pl': _('Mitglieder'),
             'depot': _('Depot'),
-            'depot_pl': _('Depots')
+            'depot_pl': _('Depots'),
+            'package': _('Tasche'),
         }
     )
     organisation_name = _get_setting('ORGANISATION_NAME', 'Juntagrico')
@@ -75,7 +76,6 @@ class Config:
     promoted_job_types = _get_setting('PROMOTED_JOB_TYPES', [])
     promomted_jobs_amount = _get_setting('PROMOTED_JOBS_AMOUNT', 2)
     depot_list_generation_days = _get_setting('DEPOT_LIST_GENERATION_DAYS', [0, 1, 2, 3, 4, 5, 6])
-    billing = _get_setting('BILLING', False)
     business_year_start = _get_setting('BUSINESS_YEAR_START', {'day': 1, 'month': 1})
     business_year_cancelation_month = _get_setting('BUSINESS_YEAR_CANCELATION_MONTH', 12)
     membership_end_month = _get_setting('MEMBERSHIP_END_MONTH', 6)