+++ juntagrico/entity/__init__.py
@@ -1,3 +1,4 @@
+from django.core.exceptions import ValidationError
 from django.db import models
 from django.utils import timezone
 from django.utils.translation import gettext as _
@@ -50,12 +51,25 @@ class SimpleStateModel(models.Model):
 
     def deactivate(self, time=None):
         now = time or timezone.now().date()
+        self.cancellation_date = self.cancellation_date or now
         self.deactivation_date = self.deactivation_date or now
         self.save()
 
     @property
+    def waiting(self):
+        return self.state == 'waiting'
+
+    @property
     def active(self):
         return self.state == 'active'
+
+    @property
+    def canceled(self):
+        return self.state == 'canceled'
+
+    @property
+    def inactive(self):
+        return self.state == 'inactive'
 
     @property
     def __state_code(self):
@@ -72,6 +86,23 @@ class SimpleStateModel(models.Model):
     @property
     def state_text(self):
         return SimpleStateModel.__state_text_dict.get(self.__state_code, _('Fehler!'))
+
+    def check_date_order(self):
+        now = timezone.now().date()
+        is_active = self.activation_date is not None
+        is_cancelled = self.cancellation_date is not None
+        is_deactivated = self.deactivation_date is not None
+        activation_date = self.activation_date or now
+        cancellation_date = self.cancellation_date or activation_date  #allow future activation date
+        deactivation_date = self.deactivation_date or cancellation_date
+        if(is_cancelled or is_deactivated) and not is_active:
+            raise ValidationError(_('Bitte "Aktivierungsdatum" ausfüllen'), code='invalid')
+        if is_deactivated and not is_cancelled:
+            raise ValidationError(_('Bitte "Kündigungsdatum" ausfüllen'), code='invalid')
+        if is_cancelled and cancellation_date > now:
+            raise ValidationError(_('Das "Kündigungsdatum" kann nicht in der Zukunft liegen'), code='invalid')
+        if not (activation_date <= cancellation_date <= deactivation_date):
+            raise ValidationError(_('Datenreihenfolge stimmt nicht.'), code='invalid')
 
     class Meta:
         abstract = True