+++ juntagrico/entity/subs.py
@@ -2,6 +2,7 @@ import datetime
 import time
 
 from django.db import models
+from django.db.models import Q
 from django.utils import timezone
 from django.utils.translation import gettext as _
 
@@ -10,9 +11,12 @@ from juntagrico.dao.sharedao import ShareDao
 from juntagrico.entity import notifiable, JuntagricoBaseModel, SimpleStateModel
 from juntagrico.entity.billing import Billable
 from juntagrico.entity.depot import Depot
+from juntagrico.entity.member import q_left_subscription, q_joined_subscription
 from juntagrico.lifecycle.sub import check_sub_consistency
-from juntagrico.util.models import q_activated, q_cancelled, q_deactivated
-from juntagrico.util.temporal import start_of_next_business_year
+from juntagrico.lifecycle.subpart import check_sub_part_consistency
+from juntagrico.util.models import q_activated, q_cancelled, q_deactivated, q_deactivation_planned, q_isactive
+from juntagrico.util.temporal import start_of_next_business_year, start_of_business_year, \
+    calculate_remaining_days_percentage
 
 
 class Subscription(Billable, SimpleStateModel):
@@ -28,6 +32,9 @@ class Subscription(Billable, SimpleStateModel):
     primary_member = models.ForeignKey('Member', related_name='subscription_primary', null=True, blank=True,
                                        on_delete=models.PROTECT,
                                        verbose_name=_('Haupt-{}-BezieherIn').format(Config.vocabulary('subscription')))
+    nickname = models.CharField(_('Spitzname'), max_length=30, blank=True,
+                                help_text=_('Ersetzt die Mit-{}-BezieherInnen auf der {}-Liste.'.format(
+                                    Config.vocabulary('subscription'), Config.vocabulary('depot'))))
     start_date = models.DateField(
         _('Gewünschtes Startdatum'), null=False, default=start_of_next_business_year)
     end_date = models.DateField(
@@ -48,34 +55,39 @@ class Subscription(Billable, SimpleStateModel):
         return '%s' % (' + '.join(namelist))
 
     @staticmethod
-    def get_size_name(parts):
-        size_dict = {}
+    def get_part_overview(parts):
+        # building multi-dimensional dictionary [product_name][size_long_name][(type_name, type_long_name)] -> amount
+        result = {}
         for part in parts.all():
-            size_dict[str(part.type)] = 1 + size_dict.get(str(part.type), 0)
-        size_names = [key + ':' + str(value) for key, value in size_dict.items()]
-        if len(size_names) > 0:
-            return '<br>'.join(size_names)
-        return _('keine {0}-Bestandteile').format(Config.vocabulary('subscription'))
+            product_name = part.type.size.product.name
+            product = result.get(product_name, {})
+            size_name = part.type.size.long_name
+            size = product.get(size_name, {})
+            type_name = (part.type.name, part.type.long_name)
+            size[type_name] = 1 + size.get(type_name, 0)
+            product[size_name] = size
+            result[product_name] = product
+        return result
 
     @property
-    def size_name(self):
-        return Subscription.get_size_name(self.active_parts)
+    def part_overview(self):
+        return Subscription.get_part_overview(self.active_parts)
 
     @property
-    def future_size_name(self):
-        return Subscription.get_size_name(self.future_parts)
+    def future_part_overview(self):
+        return Subscription.get_part_overview(self.future_parts)
 
     @property
     def active_parts(self):
-        return self.parts.filter(q_activated & ~q_deactivated)
+        return self.parts.filter(q_isactive())
 
     @property
     def future_parts(self):
-        return self.parts.filter(~q_cancelled & ~q_deactivated)
+        return self.parts.filter(~q_cancelled() & ~q_deactivated())
 
     @property
     def active_and_future_parts(self):
-        return self.parts.filter(~q_deactivated)
+        return self.parts.filter(~q_deactivated())
 
     @property
     def part_change_date(self):
@@ -87,13 +99,13 @@ class Subscription(Billable, SimpleStateModel):
     @property
     def size(self):
         sizes = {}
-        for part in self.active_parts.all():
+        for part in self.active_parts.all() or self.future_parts.all():
             sizes[part.type.size.product.name] = part.type.size.units + sizes.get(part.type.size.product.name, 0)
         return ', '.join([key + ':' + str(value) for key, value in sizes.items()])
 
     @property
     def types_changed(self):
-        return self.parts.filter(~q_activated | (q_cancelled & ~q_deactivated)).count()
+        return self.parts.filter(~q_activated() | (q_cancelled() & ~q_deactivation_planned())).count()
 
     @staticmethod
     def calc_subscription_amount(parts, size):
@@ -113,6 +125,8 @@ class Subscription(Billable, SimpleStateModel):
         result = 0
         for part in self.active_parts.all():
             result += part.type.required_assignments
+        if self.activation_date is not None and self.activation_date > start_of_business_year():
+            result = round(result * calculate_remaining_days_percentage(self.activation_date))
         return result
 
     @property
@@ -120,6 +134,8 @@ class Subscription(Billable, SimpleStateModel):
         result = 0
         for part in self.active_parts.all():
             result += part.type.required_core_assignments
+        if self.activation_date is not None and self.activation_date > start_of_business_year():
+            result = round(result * calculate_remaining_days_percentage(self.activation_date))
         return result
 
     @property
@@ -154,28 +170,48 @@ class Subscription(Billable, SimpleStateModel):
 
     recipients_names.short_description = '{}-BezieherInnen'.format(Config.vocabulary('subscription'))
 
-    def other_recipients(self):
-        return self.recipients.exclude(email=self.primary_member.email)
+    def co_members(self, member):
+        return [m.member for m in
+                self.recipients_qs.exclude(member__email=member.email).all()]
 
+    def recipients_display_name(self):
+        if self.nickname:
+            return ', '.join([self.primary_member_nullsave(), self.nickname])
+        else:
+            return '{}, {}'.format(self.primary_member_nullsave(), self.other_recipients_names)
+
+    def other_recipients(self):
+        return self.co_members(self.primary_member)
+
+    @property
     def other_recipients_names(self):
         members = self.other_recipients()
         return ', '.join(str(member) for member in members)
 
     @property
+    def recipients_qs(self):
+        return self.memberships_for_state.order_by(
+            'member__first_name', 'member__last_name')
+
+    @property
     def recipients(self):
-        return self.recipients_all.filter(inactive=False)
+        return [m.member for m in self.recipients_qs.all()]
 
     @property
     def recipients_all(self):
-        return self.recipients_all_for_state(self.state)
+        return [m.member for m in self.memberships_for_state.all()]
 
-    def recipients_all_for_state(self, state):
-        if state == 'waiting':
-            return self.members_future.all()
-        elif state == 'inactive':
-            return self.members_old.all()
+    @property
+    def memberships_for_state(self):
+        now = timezone.now().date()
+        member_active = ~Q(member__deactivation_date__isnull=False, member__deactivation_date__lte=now)
+        if self.state == 'waiting':
+            return self.subscriptionmembership_set.prefetch_related('member').filter(member_active)
+        elif self.state == 'inactive':
+            return self.subscriptionmembership_set.prefetch_related('member')
         else:
-            return self.members.all()
+            return self.subscriptionmembership_set.filter(q_joined_subscription(),
+                                                          ~q_left_subscription(), member_active).prefetch_related('member')
 
     def primary_member_nullsave(self):
         member = self.primary_member
@@ -185,11 +221,11 @@ class Subscription(Billable, SimpleStateModel):
 
     @property
     def extra_subscriptions(self):
-        return self.extra_subscription_set.filter(q_activated & ~q_deactivated)
+        return self.extra_subscription_set.filter(q_isactive())
 
     @property
     def future_extra_subscriptions(self):
-        return self.extra_subscription_set.filter(~q_cancelled & ~q_deactivated)
+        return self.extra_subscription_set.filter(~q_cancelled() & ~q_deactivated())
 
     @property
     def extrasubscriptions_changed(self):
@@ -222,7 +258,8 @@ class Subscription(Billable, SimpleStateModel):
     class Meta:
         verbose_name = Config.vocabulary('subscription')
         verbose_name_plural = Config.vocabulary('subscription_pl')
-        permissions = (('can_filter_subscriptions', _('Benutzer kann {0} filtern').format(Config.vocabulary('subscription'))),)
+        permissions = (
+            ('can_filter_subscriptions', _('Benutzer kann {0} filtern').format(Config.vocabulary('subscription'))),)
 
 
 class SubscriptionPart(JuntagricoBaseModel, SimpleStateModel):
@@ -233,9 +270,12 @@ class SubscriptionPart(JuntagricoBaseModel, SimpleStateModel):
 
     @property
     def can_cancel(self):
-        # TODO
-        return True
+        return self.cancellation_date is None and self.subscription.future_parts.count() > 1
 
+    def clean(self):
+        check_sub_part_consistency(self)
+
+    @notifiable
     class Meta:
         verbose_name = _('{} Bestandteil').format(Config.vocabulary('subscription'))
         verbose_name_plural = _('{} Bestandteile').format(Config.vocabulary('subscription'))