+++ juntagrico/views.py
@@ -3,6 +3,7 @@ from django.contrib.auth.decorators import login_required
 from django.db.models import Count
 from django.http import HttpResponse
 from django.shortcuts import render, get_object_or_404, redirect
+from django.urls import reverse
 from django.utils import timezone
 from django.utils.translation import gettext as _
 
@@ -42,9 +43,9 @@ def get_menu_dict(request):
     next_jobs = JobDao.upcomming_jobs_for_member(member)
 
     required_assignments = 0
-    if member.subscription is not None:
+    if member.subscription_current is not None:
         partner_assignments = []
-        for subscription_member in member.subscription.recipients_all:
+        for subscription_member in member.subscription_current.recipients_all:
             if subscription_member == member:
                 continue
             partner_assignments.extend(
@@ -52,7 +53,7 @@ def get_menu_dict(request):
 
         userassignments = AssignmentDao.assignments_for_member_current_business_year(
             member)
-        required_assignments = member.subscription.required_assignments
+        required_assignments = member.subscription_current.required_assignments
     else:
         partner_assignments = []
         userassignments = []
@@ -86,7 +87,7 @@ def get_menu_dict(request):
         'area_admin': area_admin,
         'show_core': ActivityAreaDao.all_core_areas().count() > 0,
         'show_extras': JobExtraDao.all_job_extras().count() > 0,
-        'show_deliveries': len(DeliveryDao.deliveries_by_subscription(request.user.member.subscription)) > 0,
+        'show_deliveries': len(DeliveryDao.deliveries_by_subscription(request.user.member.subscription_current)) > 0,
         'admin_menus': addons.config.get_admin_menus(),
         'admin_subscription_menus': addons.config.get_admin_subscription_menu(),
         'user_menus': addons.config.get_user_menus(),
@@ -168,7 +169,7 @@ def job(request, job_id):
             name += _(' (mit einer weiteren Person)')
         elif participant.assignment_for_job > 2:
             name += _(' (mit {} weiteren Personen)').format(participant.assignment_for_job - 1)
-        contact_url = '/my/contact/member/{}/{}/'.format(participant.id, job_id)
+        contact_url = reverse('contact-member', args=[participant.id])
         extras = []
         for assignment in AssignmentDao.assignments_for_job_and_member(job.id, participant):
             for extra in assignment.job_extras.all():
@@ -215,7 +216,7 @@ def areas(request):
     '''
     member = request.user.member
     my_areas = []
-    for area in ActivityAreaDao.all_visible_areas():
+    for area in ActivityAreaDao.all_visible_areas_ordered():
         my_areas.append({
             'name': area.name,
             'checked': member in area.members.all(),
@@ -330,7 +331,7 @@ def deliveries(request):
     '''
     renderdict = get_menu_dict(request)
     deliveries = DeliveryDao.deliveries_by_subscription(
-        request.user.member.subscription)
+        request.user.member.subscription_current)
     renderdict.update({
         'deliveries': deliveries,
         'menu': {'deliveries': 'active'},
@@ -362,30 +363,31 @@ def contact(request):
 
 
 @login_required
-def contact_member(request, member_id, job_id):
+def contact_member(request, member_id):
     '''
     member contact form
     '''
     member = request.user.member
     contact_member = get_object_or_404(Member, id=int(member_id))
     is_sent = False
+    back_url = request.META.get('HTTP_REFERER') or reverse('home')
 
     if request.method == 'POST':
         # send mail to member
+        back_url = request.POST.get('back_url')
         files = []
         append_attachements(request, files)
         formemails.contact_member(request.POST.get('subject'), request.POST.get('message'), member, contact_member,
                                   request.POST.get('copy'), files)
         is_sent = True
-    job = JobDao.job_by_id(job_id)
     renderdict = get_menu_dict(request)
     renderdict.update({
-        'admin': request.user.is_staff or job.type.activityarea.coordinator == member,
+        'admin': request.user.has_perm('juntagrico.is_operations_group') or request.user.has_perm('juntagrico.is_area_admin'),
         'usernameAndEmail': member.first_name + ' ' + member.last_name + '<' + member.email + '>',
         'member_id': member_id,
         'member_name': contact_member.first_name + ' ' + contact_member.last_name,
         'is_sent': is_sent,
-        'job_id': job_id
+        'back_url': back_url
     })
     return render(request, 'contact_member.html', renderdict)
 
@@ -430,13 +432,12 @@ def cancel_membership(request):
         end_date = next_membership_end_date()
         message = request.POST.get('message')
         member = request.user.member
-        member.canceled = True
         member.end_date = end_date
         member.cancelation_date = now
         if member.is_cooperation_member:
             adminnotification.member_canceled(member, end_date, message)
         else:
-            member.inactive = True
+            member.deactivation_date = now
 
         member.save()
         for share in member.active_shares:
@@ -446,8 +447,8 @@ def cancel_membership(request):
     missing_iban = member.iban == ''
     coop_member = member.is_cooperation_member
     asc = member.usable_shares_count
-    sub = member.subscription
-    f_sub = member.future_subscription
+    sub = member.subscription_current
+    f_sub = member.subscription_future
     future_active = f_sub is not None and f_sub.state == 'active' and f_sub.state == 'waiting'
     current_active = sub is not None and sub.state == 'active' and sub.state == 'waiting'
     future = future_active and f_sub.share_overflow - asc < 0